// store/product.type.ts
export interface Product {
    id: number;
    name: string;
    power: string;
    quantity: number;
    img_url: string;
    price: number;
    description: string;
    brand: string;
    weight: number;
    height: number;
    width: number;
    length: number;
    colour: string;
    model_code: string;
}